feat(app): unify error handling (#431)
