+++ src/containers/user-profile-form-modal.container.tsx
@@ -18,7 +18,6 @@ import {
   optional,
   url,
 } from 'development-kit/form';
-import { parseMessage } from 'development-kit/parse-message';
 import { useFileInput } from 'development-kit/use-file-input';
 import { useForm } from 'development-kit/use-form';
 import { useToggle } from 'development-kit/use-toggle';
@@ -77,43 +76,6 @@ const validators: ValidatorsSetup<UserProfileFormValues> = {
   linkedInUrl: urlValidator,
   fbUrl: urlValidator,
   twitterUrl: urlValidator,
-};
-
-const UpdateErrorModal = ({
-  onSync,
-}: {
-  onSync: UserProfileFormModalContainerProps['onSync'];
-}) => {
-  const updateYourProfileStore = updateYourProfileStoreSelectors.useState();
-
-  if (updateYourProfileStore.is !== `fail`) return null;
-
-  const parsed = parseMessage(updateYourProfileStore.error);
-
-  return (
-    <ErrorModal
-      heading="Ups, something went wrong"
-      message={parsed.message}
-      footer={
-        parsed.symbol === `outOfDateEntry` && (
-          <Button
-            type="button"
-            i={2}
-            s={2}
-            auto
-            title="Sync Your profile"
-            onClick={() => {
-              updateYourProfileStoreActions.idle();
-              onSync();
-            }}
-          >
-            Sync
-          </Button>
-        )
-      }
-      onClose={updateYourProfileStoreActions.idle}
-    />
-  );
 };
 
 const UserProfileFormModalContainer = ({
@@ -328,7 +290,30 @@ const UserProfileFormModalContainer = ({
         <Status>Updating your profile...</Status>
       )}
 
-      <UpdateErrorModal onSync={onSync} />
+      {updateYourProfileStore.is === `fail` && (
+        <ErrorModal
+          heading="Ups, something went wrong"
+          message={updateYourProfileStore.error.message}
+          footer={
+            updateYourProfileStore.error.symbol === `out-of-date` && (
+              <Button
+                type="button"
+                i={2}
+                s={2}
+                auto
+                title="Sync Your profile"
+                onClick={() => {
+                  updateYourProfileStoreActions.idle();
+                  onSync();
+                }}
+              >
+                Sync
+              </Button>
+            )
+          }
+          onClose={updateYourProfileStoreActions.idle}
+        />
+      )}
 
       {avatarErrorModal.opened && (
         <ErrorModal
