+++ src/development-kit/parse-error.ts
@@ -1,9 +1,96 @@
-const parseError = (error: unknown): string => {
-  if (error instanceof Error) {
-    return error.message;
-  }
+const symbols = [
+  `already-exists`,
+  `unauthenticated`,
+  `internal`,
+  `invalid-schema`,
+  `not-found`,
+  `out-of-date`,
+  `bad-request`,
+] as const;
 
-  return `Ups, something went wrong. Try again.`;
+type ErrorSymbol = (typeof symbols)[number];
+type ErrorContent = string | { key: string; message: string }[];
+
+type ErrorVariant<
+  TSymbol extends ErrorSymbol,
+  TContent extends ErrorContent = string,
+> = {
+  symbol: TSymbol;
+  content: TContent;
 };
 
+type AlreadyExistsError = ErrorVariant<`already-exists`>;
+type UnauthenticatedError = ErrorVariant<`unauthenticated`>;
+type InternalError = ErrorVariant<`internal`>;
+type InvalidSchemaError = ErrorVariant<
+  `invalid-schema`,
+  { key: string; message: string }[]
+>;
+type NotFoundError = ErrorVariant<`not-found`>;
+type OutOfDateError = ErrorVariant<`out-of-date`>;
+type BadRequestError = ErrorVariant<`bad-request`>;
+
+type KnownError =
+  | AlreadyExistsError
+  | UnauthenticatedError
+  | InternalError
+  | InvalidSchemaError
+  | NotFoundError
+  | OutOfDateError
+  | BadRequestError;
+
+type ParsedError = (
+  | KnownError
+  | {
+      symbol: 'unknown';
+      content: string;
+    }
+) & { message: string };
+
+const parseError = (error: unknown): ParsedError => {
+  const unknownError: ParsedError = {
+    symbol: `unknown`,
+    content: `Unknown error occured`,
+    message: `Unknown error occured`,
+  };
+
+  if (typeof error !== `string`) {
+    return unknownError;
+  }
+
+  try {
+    const { symbol, content } = JSON.parse(error) as KnownError;
+
+    if (!content || !symbol) {
+      return unknownError;
+    }
+
+    if (!symbols.includes(symbol)) {
+      return unknownError;
+    }
+
+    if (
+      symbol === `invalid-schema` &&
+      (!Array.isArray(content) ||
+        content.length === 0 ||
+        content.some(
+          (reason) =>
+            typeof reason.key !== `string` ||
+            typeof reason.message !== `string`,
+        ))
+    ) {
+      return unknownError;
+    }
+    // @TODO[PRIO=5]: [Think about error shape design that does not need multiple models].
+    return {
+      symbol,
+      content,
+      message: Array.isArray(content) ? content[0].message : content,
+    } as ParsedError;
+  } catch {
+    return unknownError;
+  }
+};
+
+export type { ParsedError };
 export { parseError };
