+++ src/features/creator/containers/creator-error-modal.container.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ErrorModal from 'components/error-modal';
 import { Button } from 'design-system/button';
-import { parseMessage } from 'development-kit/parse-message';
 import { authStoreSelectors } from 'store/auth/auth.store';
 import {
   docManagementStoreActions,
@@ -11,15 +10,13 @@ import {
 const CreatorErrorModalContainer = () => {
   const docManagementStore = docManagementStoreSelectors.useFail();
   const authStore = authStoreSelectors.useAuthorized();
-  const parsed = parseMessage(docManagementStore.error);
 
   return (
     <ErrorModal
       heading="Ups, something went wrong"
-      message={parsed.message}
+      message={docManagementStore.error.message}
       footer={
-        (parsed.symbol === `outOfDateEntry` ||
-          parsed.symbol === `out-of-date`) && (
+        docManagementStore.error.symbol === `out-of-date` && (
           <Button
             type="button"
             i={2}
